+++ src/main/java/NodeK.java
@@ -1,17 +1,18 @@
 
 
 class NodeK {
-	int attr1;
-	int attr2;
+	Double attr1;
+	Double attr2;
 	int label;
 	int label2;
-
-	public NodeK(int attr1, int attr2, int label,int label2) {
+	public NodeK(Double attr1, Double attr2, int label, int label2) {
 		super();
 		this.attr1 = attr1;
 		this.attr2 = attr2;
 		this.label = label;
 		this.label2 = label2;
 	}
+
+	
 
 }
